PDL-4231: expose internal error and mapped error 9003
